Sanitize toegevoegd
Input schoonmaak functie
