+++ core.php
@@ -28,7 +28,8 @@ class Core {
         }
         return $language;
     }	
-//Deze functie is om de juiste pagina weer te geven uit de database
+	
+	//Deze functie is om de juiste pagina weer te geven uit de database
     function page($page) {
         if (!file_exists('page/' . $page . '.php')) {
             include 'page/home.php';
@@ -36,6 +37,7 @@ class Core {
             include 'page/' . $page . '.php';
         }
     }
+	
     //hiervandaan wordt het portfolio weergegeven
     function getPortfolio($id) {
         $dbc = $this->dbc();
@@ -60,11 +62,23 @@ class Core {
         
         mysqli_query($dbc, $sql) or die("De pagina kan niet worden aangemaakt");
     }
+	
     //Hier wordt het portfolio geupdate
     function editPortfolio($userID, $content, $tags, $ContentID){
         $dbc = $this->dbc();
         $sql= "UPDATE `content` SET `Content` = '" . $content . "', Tags='" . $tags . "' WHERE `content`.`ContentID` = " . $ContentID .";";
          mysqli_query($dbc, $sql) or die("De pagina kan niet worden aangemaakt");
     }
+	
+	// Input velden opruimen van code en andere html tekens
+	function Sanitize($input){
+		require("inc/conn.php");
+		
+		$input 		= stripslashes($input);
+		$input		= mysqli_real_escape_string($conn, $input);
+		$input		= htmlentities($input);
+		
+		return $input;
+	}
 }
 ?>
